move gamefield to file
